+++ app/src/main/java/com/king/ultraswiperefresh/app/sample/LottieRefreshIndicatorSample.kt
@@ -1,8 +1,6 @@
 package com.king.ultraswiperefresh.app.sample
 
 import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.material3.HorizontalDivider
@@ -15,7 +13,6 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
@@ -23,7 +20,6 @@ import androidx.compose.ui.unit.sp
 import com.king.ultraswiperefresh.NestedScrollMode
 import com.king.ultraswiperefresh.UltraSwipeRefresh
 import com.king.ultraswiperefresh.app.component.ColumnItem
-import com.king.ultraswiperefresh.indicator.classic.ClassicRefreshFooter
 import com.king.ultraswiperefresh.indicator.lottie.LottieRefreshFooter
 import com.king.ultraswiperefresh.indicator.lottie.LottieRefreshHeader
 import com.king.ultraswiperefresh.rememberUltraSwipeRefreshState